pyproject.toml
[project]
name = "hvac-mcp"
version = "0.1.0"
description = "Model Context Protocol server for HVAC and plumbing technicians"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Kollin Croyle", email = "cybertec44@yahoo.com" }
]
keywords = ["mcp", "hvac", "plumbing", "claude", "field-service", "refrigerant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
hvac-mcp = "hvac_mcp.server:main"

[project.urls]
Homepage = "https://hvac-mcp.nightfast.tech"
Repository = "https://github.com/cybertec44/hvac-mcp"
Issues = "https://github.com/cybertec44/hvac-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hvac_mcp"]

[tool.hatch.build.targets.wheel.force-include]
"src/hvac_mcp/data" = "hvac_mcp/data"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
